Separate CLI code into a separate submodule
Makes clear the distinction between the main package python API and
the CLI which consumes it.
